+++ test/test-buildDocumentFile/src/home3/thed.tpl.js
@@ -1,2 +1,0 @@
-// returns false if props is falsy, or the normal promise if not.
-module.exports = (mojl, props) => !!props && mojl.template`>>>${props.foo}<<<`;